Update more examples to align with the recommended path
